don't try and delete a empty list of files.
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/DIRAC/trunk/DIRAC@21223 4525493e-7705-40b1-a816-d608a930855b

+++ TransformationSystem/DB/TransformationDB.py
@@ -1366,7 +1366,8 @@ class TransformationDB(DB):
       for fileID in fileIDs.keys():
         if not fileID in fileReplicas.keys():
           noReplicas.append(fileIDs[fileID])
-      self.removeFile(noReplicas)
+      if noReplicas:
+        self.removeFile(noReplicas)
     resDict = {'Successful':successful,'Failed':failed}
     return S_OK(resDict)
 
@@ -1377,6 +1378,8 @@ class TransformationDB(DB):
     failed = {}
     successful = {}
     connection = self.__getConnection(connection)
+    if not lfns:
+      return S_ERROR("No LFNs supplied")
     res = self.__getFileIDsForLfns(lfns, connection=connection)
     if not res['OK']:
       return res
